[Adapt]Added support for get adapter property APIs
This patch added support to get adapter property APIs mentioned below:
1/ _bt_get_local_address
2/ _bt_get_local_version
3/ _bt_get_local_name
4/ _bt_get_discoverable_mode

+++ bt-oal/common/oal-event-dispatcher.c
@@ -34,7 +34,7 @@
 
 typedef struct {
 	int event;
-	unsigned int size;
+	gsize size;
 	gpointer event_data;
 } event_t;
